+++ WebServer/dispatcher.py
@@ -7,10 +7,9 @@ import os.path
 import mimetypes
 from datetime import datetime
 
-from WebServer.tools import setupHeadersResponse
+from WebServer.headerResponse import setupHeadersResponse
+
 from WebServer.logging import logging
-
-
 from WebServer.rest_Bindings import rest_bindLSTcluster, rest_bindLSTdevice, rest_binding, rest_unbinding
 from WebServer.rest_Topology import rest_netTopologie, rest_req_topologie
 from WebServer.rest_Energy import rest_req_nwk_full, rest_req_nwk_inter
