+++ WebServer/rest_Topology.py
@@ -10,7 +10,7 @@ from datetime import datetime
 
 
 
-from WebServer.tools import setupHeadersResponse
+from WebServer.headerResponse import setupHeadersResponse
 
 def rest_req_topologie( self, verb, data, parameters):
     _response = setupHeadersResponse()
